refactor: extract query modifier logic out of the Subscription scope
