+++ packages/jazz-tools/src/tools/internal.ts
@@ -28,6 +28,7 @@ export * from "./implementation/schema.js";
 export * from "./subscribe/SubscriptionScope.js";
 export * from "./subscribe/types.js";
 export * from "./subscribe/index.js";
+export * from "./subscribe/queryModifiers.js";
 export * from "./lib/cache.js";
 export * from "./lib/utilityTypes.js";
 export * from "./lib/utils.js";
