+++ packages/jazz-tools/src/tools/subscribe/SubscriptionScope.ts
@@ -24,47 +24,11 @@ import type {
   SubscriptionValue,
 } from "./types.js";
 import { createCoValue, myRoleForRawValue } from "./utils.js";
-
-export const OrderByDirection = { ASC: "asc", DESC: "desc" } as const;
-export type OrderByDirection =
-  (typeof OrderByDirection)[keyof typeof OrderByDirection];
-export const WhereComparisonOperators = {
-  $eq: "$eq",
-  $ne: "$ne",
-  $gt: "$gt",
-  $gte: "$gte",
-  $lt: "$lt",
-  $lte: "$lte",
-} as const;
-export type WhereComparisonOperator = keyof typeof WhereComparisonOperators;
-
-export const WhereLogicalOperators = {
-  $and: "$and",
-  $or: "$or",
-  $not: "$not",
-} as const;
-export type WhereLogicalOperator = keyof typeof WhereLogicalOperators;
-
-export type WhereClause =
-  | {
-      field: string;
-      operator: WhereComparisonOperator;
-      value: any;
-    }
-  | {
-      combinator: WhereLogicalOperator;
-      conditions: WhereClause[];
-    };
-
-type QueryModifiers = {
-  where?: WhereClause;
-  orderBy?: {
-    field: string;
-    orderDirection: OrderByDirection;
-  }[];
-  limit?: number;
-  offset?: number;
-};
+import {
+  parseQueryModifiers,
+  queryModifierFields,
+  QueryModifiers,
+} from "./queryModifiers.js";
 
 export class SubscriptionScope<D extends CoValue> {
   childNodes = new Map<string, SubscriptionScope<CoValue>>();
@@ -111,29 +75,8 @@ export class SubscriptionScope<D extends CoValue> {
     this.resolve = resolve;
     this.value = { type: "unloaded", id };
 
-    const queryModifiers: Record<string, any> =
-      typeof this.resolve === "object" && this.resolve !== null
-        ? pick(this.resolve, [
-            "$where",
-            "$orderBy",
-            "$limit",
-            "$offset",
-          ] as const)
-        : {};
-    if (Object.keys(queryModifiers).length > 0) {
-      this.queryModifiers.limit = queryModifiers?.$limit;
-      this.queryModifiers.offset = queryModifiers?.$offset;
-      this.queryModifiers.where = parseWhere(queryModifiers?.$where);
-      this.queryModifiers.orderBy = parseOrderBy(queryModifiers?.$orderBy);
-    } else {
-      this.queryModifiers = {};
-    }
-    const indexedFields = [
-      ...(this.queryModifiers.orderBy?.map((orderBy) => orderBy.field) ?? []),
-      ...(this.queryModifiers?.where
-        ? indexedFieldsFromWhereClause(this.queryModifiers?.where)
-        : []),
-    ].filter((field): field is string => field !== undefined);
+    this.queryModifiers = parseQueryModifiers(this.resolve);
+    const queryFields = queryModifierFields(this.queryModifiers);
 
     let lastUpdate: RawCoValue | "unavailable" | undefined;
     this.subscription = new CoValueCoreSubscription(
@@ -148,8 +91,7 @@ export class SubscriptionScope<D extends CoValue> {
           return;
         }
 
-        // TODO rename variable
-        if (indexedFields.length > 0 && value !== "unavailable") {
+        if (queryFields.length > 0 && value !== "unavailable") {
           this.requestCoListChildrenLoad();
         }
 
@@ -784,124 +726,4 @@ export class SubscriptionScope<D extends CoValue> {
     this.subscribers.clear();
     this.childNodes.forEach((child) => child.destroy());
   }
-}
-
-function parseWhere(where: any): QueryModifiers["where"] {
-  if (!where) return undefined;
-
-  const topLevelKeys = Object.keys(where);
-
-  const logicalOperatorKeys = topLevelKeys.filter(
-    (key) => key in WhereLogicalOperators,
-  );
-  const logicalConditions = logicalOperatorKeys.map((key) => {
-    const combinator = key as WhereLogicalOperator;
-    const conditions = where[combinator];
-
-    if (combinator === WhereLogicalOperators.$not) {
-      const negatedCondition = parseWhere(conditions as WhereOptions<any>);
-      return {
-        combinator,
-        conditions: negatedCondition ? [negatedCondition] : [],
-      };
-    } else {
-      return {
-        combinator,
-        conditions: (conditions as WhereOptions<any>[])
-          .map((cond) => parseWhere(cond))
-          .filter((cond) => cond !== undefined),
-      };
-    }
-  });
-
-  const fieldConditionsKeys = topLevelKeys.filter(
-    (key) => !(key in WhereLogicalOperators),
-  );
-  const fieldConditions = Object.entries(
-    pick(where, fieldConditionsKeys),
-  ).flatMap(([field, filter]) => parseFieldConditions(field, filter));
-
-  const allConditions = [...logicalConditions, ...fieldConditions];
-
-  if (allConditions.length === 1) {
-    return allConditions[0];
-  }
-  return {
-    combinator: WhereLogicalOperators.$and,
-    conditions: allConditions,
-  };
-}
-
-function parseFieldConditions(
-  field: string,
-  filter: WhereFieldCondition<any>,
-): WhereClause {
-  if (typeof filter !== "object") {
-    return {
-      field,
-      operator: WhereComparisonOperators.$eq,
-      value: filter,
-    };
-  }
-
-  const filterKeys = Object.keys(filter);
-
-  const logicalFilters = filterKeys
-    .filter((key) => key in WhereLogicalOperators)
-    .map((key) => {
-      const logicalOperator = key as WhereLogicalOperator;
-      const conditions = filter[logicalOperator];
-
-      if (logicalOperator === WhereLogicalOperators.$not) {
-        return {
-          combinator: logicalOperator,
-          conditions: [parseFieldConditions(field, conditions)],
-        };
-      } else {
-        const fieldConditions = conditions.map((cond: any) =>
-          parseFieldConditions(field, cond),
-        );
-        return {
-          combinator: logicalOperator,
-          conditions: fieldConditions,
-        };
-      }
-    });
-
-  const comparisonFilters = filterKeys
-    .filter((key) => key in WhereComparisonOperators)
-    .map((key) => ({
-      field,
-      operator: key as WhereComparisonOperator,
-      value: filter[key],
-    }));
-
-  const allFilters = [...logicalFilters, ...comparisonFilters];
-  if (allFilters.length === 1) {
-    return allFilters[0]!;
-  }
-  return {
-    combinator: WhereLogicalOperators.$and,
-    conditions: allFilters,
-  };
-}
-
-function indexedFieldsFromWhereClause(where: WhereClause): string[] {
-  if ("field" in where) {
-    return [where.field];
-  } else {
-    return where.conditions.flatMap((condition) =>
-      indexedFieldsFromWhereClause(condition),
-    );
-  }
-}
-
-function parseOrderBy(orderBy: any): QueryModifiers["orderBy"] {
-  const orderClauses = Object.entries(orderBy ?? {});
-  return orderClauses.map(([field, orderDirection]) => {
-    return {
-      field,
-      orderDirection: orderDirection as OrderByDirection,
-    };
-  });
 }
